pyproject.toml
[tool.isort]
profile = "black"

[tool.poetry]
name = "pycurrentwindow"
version = "0.1.0"
description = ""
authors = ["David <45048997+astridos2go@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^9.3.0"
pystray = "^0.19.4"
psutil = "^5.9.4"
pyserial = "^3.5"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
flake8 = "^6.0.0"
mypy = "^0.991"
isort = "^5.10.1"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
